requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
PySocks==1.7.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1

#for telegram
python-telegram-bot[socks]==13.4.1

# for web
requests==2.25.1
chardet==4.0.0
idna==2.10
PySocks==1.7.1
urllib3==1.26.4

# for database
sqlalchemy==1.4.6